Try to restore build on build server.
The tests in image pyramid are failing on build.geoserver.org. The test that's failing is in a base class, and ends up running multiple times. One of them, it fails. Refactored the code so that it gets run only once instead.
